Refactor: use namespaces for api versions
