+++ plugins/orchestrator-backend/src/service/api/v1.ts
@@ -14,126 +14,129 @@ import {
 import { DataIndexService } from '../DataIndexService';
 import { SonataFlowService } from '../SonataFlowService';
 
-export async function getWorkflowOverviewV1(
-  sonataFlowService: SonataFlowService,
-): Promise<WorkflowOverviewListResult> {
-  const overviews = await sonataFlowService.fetchWorkflowOverviews();
-  if (!overviews) {
-    throw new Error("Couldn't fetch workflow overviews");
-  }
-  const result: WorkflowOverviewListResult = {
-    items: overviews,
-    limit: 0,
-    offset: 0,
-    totalCount: overviews?.length ?? 0,
-  };
-  return result;
-}
-
-export async function getWorkflowOverviewByIdV1(
-  sonataFlowService: SonataFlowService,
-  workflowId: string,
-): Promise<WorkflowOverview> {
-  const overviewObj = await sonataFlowService.fetchWorkflowOverview(workflowId);
-
-  if (!overviewObj) {
-    throw new Error(`Couldn't fetch workflow overview for ${workflowId}`);
-  }
-  return overviewObj;
-}
-
-export async function getWorkflowsV1(
-  sonataFlowService: SonataFlowService,
-  dataIndexService: DataIndexService,
-): Promise<WorkflowListResult> {
-  const definitions: WorkflowInfo[] =
-    await dataIndexService.getWorkflowDefinitions();
-  const items: WorkflowItem[] = await Promise.all(
-    definitions.map(async info => {
-      const uri = await sonataFlowService.fetchWorkflowUri(info.id);
-      if (!uri) {
-        throw new Error(`Uri is required for workflow ${info.id}`);
-      }
-      const item: WorkflowItem = {
-        definition: info as WorkflowDefinition,
-        serviceUrl: info.serviceUrl,
-        uri,
-      };
-      return item;
-    }),
-  );
-
-  if (!items) {
-    throw new Error("Couldn't fetch workflows");
+export namespace V1 {
+  export async function getWorkflowOverview(
+    sonataFlowService: SonataFlowService,
+  ): Promise<WorkflowOverviewListResult> {
+    const overviews = await sonataFlowService.fetchWorkflowOverviews();
+    if (!overviews) {
+      throw new Error("Couldn't fetch workflow overviews");
+    }
+    const result: WorkflowOverviewListResult = {
+      items: overviews,
+      limit: 0,
+      offset: 0,
+      totalCount: overviews?.length ?? 0,
+    };
+    return result;
   }
 
-  return {
-    items: items,
-    limit: 0,
-    offset: 0,
-    totalCount: items?.length ?? 0,
-  };
-}
+  export async function getWorkflowOverviewById(
+    sonataFlowService: SonataFlowService,
+    workflowId: string,
+  ): Promise<WorkflowOverview> {
+    const overviewObj =
+      await sonataFlowService.fetchWorkflowOverview(workflowId);
 
-export async function getWorkflowByIdV1(
-  sonataFlowService: SonataFlowService,
-  workflowId: string,
-): Promise<{ uri: string; definition: WorkflowDefinition }> {
-  const definition =
-    await sonataFlowService.fetchWorkflowDefinition(workflowId);
-
-  if (!definition) {
-    throw new Error(`Couldn't fetch workflow definition for ${workflowId}`);
+    if (!overviewObj) {
+      throw new Error(`Couldn't fetch workflow overview for ${workflowId}`);
+    }
+    return overviewObj;
   }
 
-  const uri = await sonataFlowService.fetchWorkflowUri(workflowId);
-  if (!uri) {
-    throw new Error(`Couldn't fetch workflow uri for ${workflowId}`);
-  }
-
-  return { uri, definition };
-}
-
-export async function getInstancesV1(
-  dataIndexService: DataIndexService,
-): Promise<ProcessInstance[]> {
-  const instances = await dataIndexService.fetchProcessInstances();
-
-  if (!instances) {
-    throw new Error("Couldn't fetch process instances");
-  }
-  return instances;
-}
-
-export async function getInstanceByIdV1(
-  dataIndexService: DataIndexService,
-  instanceId: string,
-  includeAssessment?: string,
-): Promise<AssessedProcessInstance> {
-  const instance = await dataIndexService.fetchProcessInstance(instanceId);
-
-  if (!instance) {
-    throw new Error(`Couldn't fetch process instance ${instanceId}`);
-  }
-
-  let assessedByInstance: ProcessInstance | undefined;
-
-  if (!!includeAssessment && instance.businessKey) {
-    assessedByInstance = await dataIndexService.fetchProcessInstance(
-      instance.businessKey,
+  export async function getWorkflows(
+    sonataFlowService: SonataFlowService,
+    dataIndexService: DataIndexService,
+  ): Promise<WorkflowListResult> {
+    const definitions: WorkflowInfo[] =
+      await dataIndexService.getWorkflowDefinitions();
+    const items: WorkflowItem[] = await Promise.all(
+      definitions.map(async info => {
+        const uri = await sonataFlowService.fetchWorkflowUri(info.id);
+        if (!uri) {
+          throw new Error(`Uri is required for workflow ${info.id}`);
+        }
+        const item: WorkflowItem = {
+          definition: info as WorkflowDefinition,
+          serviceUrl: info.serviceUrl,
+          uri,
+        };
+        return item;
+      }),
     );
+
+    if (!items) {
+      throw new Error("Couldn't fetch workflows");
+    }
+
+    return {
+      items: items,
+      limit: 0,
+      offset: 0,
+      totalCount: items?.length ?? 0,
+    };
   }
 
-  const response: AssessedProcessInstance = {
-    instance,
-    assessedBy: assessedByInstance,
-  };
-  return response;
-}
+  export async function getWorkflowById(
+    sonataFlowService: SonataFlowService,
+    workflowId: string,
+  ): Promise<{ uri: string; definition: WorkflowDefinition }> {
+    const definition =
+      await sonataFlowService.fetchWorkflowDefinition(workflowId);
 
-export function extractQueryParamV1(
-  req: express.Request,
-  key: string,
-): string | undefined {
-  return req.query[key] as string | undefined;
+    if (!definition) {
+      throw new Error(`Couldn't fetch workflow definition for ${workflowId}`);
+    }
+
+    const uri = await sonataFlowService.fetchWorkflowUri(workflowId);
+    if (!uri) {
+      throw new Error(`Couldn't fetch workflow uri for ${workflowId}`);
+    }
+
+    return { uri, definition };
+  }
+
+  export async function getInstances(
+    dataIndexService: DataIndexService,
+  ): Promise<ProcessInstance[]> {
+    const instances = await dataIndexService.fetchProcessInstances();
+
+    if (!instances) {
+      throw new Error("Couldn't fetch process instances");
+    }
+    return instances;
+  }
+
+  export async function getInstanceById(
+    dataIndexService: DataIndexService,
+    instanceId: string,
+    includeAssessment?: string,
+  ): Promise<AssessedProcessInstance> {
+    const instance = await dataIndexService.fetchProcessInstance(instanceId);
+
+    if (!instance) {
+      throw new Error(`Couldn't fetch process instance ${instanceId}`);
+    }
+
+    let assessedByInstance: ProcessInstance | undefined;
+
+    if (!!includeAssessment && instance.businessKey) {
+      assessedByInstance = await dataIndexService.fetchProcessInstance(
+        instance.businessKey,
+      );
+    }
+
+    const response: AssessedProcessInstance = {
+      instance,
+      assessedBy: assessedByInstance,
+    };
+    return response;
+  }
+
+  export function extractQueryParam(
+    req: express.Request,
+    key: string,
+  ): string | undefined {
+    return req.query[key] as string | undefined;
+  }
 }
